Completely rewritten (again XD) Ink now serves general purposes (not just gui)
The purpose of ink is to: create components and connect them to entities

It works as follows:

For each function executed in ink (via doComponentsFunction
("function_name")), the functions of the same name of the components are
executed (it only executes the components that contain a function with
this name).

For each function performed on a component, the entities that contain
that component (that is, a key with the name of the component file) are
taken as parameters.

Entities are tables that primarily contain id, name, tag, group, and
parent (id).

When you attach a component to an entity, the entity creates a key with
the name of the component's file with the properties that the component
will use to perform its function, and will be inserted into the
'entities' property of the component's id of that entity.